chore(grpc-asyncio): replace hatch with poetry
Since poetry is used in our other python libraries, and hatch is
hard to understand (and to get working), we replace hatch with
poetry.

As a nifty side-effect, the Docker container is now executed as
a user with the same uid/gid as the host user, resulting in
files created with the correct permissions in the mounted host
volume.